requirements.txt
streamlit
chromadb
PyPDF2
docx
pandas
openpyxl
requests
python-dotenv
google-generative-ai
supabase
psycopg2-binary
moviepy
opencv-python
openpyxl
xlrd
